Create `PreGenerateBillingDataService`
