+++ app/services/supplementary-billing/process-billing-period.service.js
@@ -16,8 +16,7 @@ const DetermineChargePeriodService = require('./determine-charge-period.service.
 const DetermineMinimumChargeService = require('./determine-minimum-charge.service.js')
 const FetchInvoiceAccountNumbersService = require('./fetch-invoice-account-numbers.service.js')
 const GenerateBillingTransactionsService = require('./generate-billing-transactions.service.js')
-const GenerateBillingInvoiceService = require('./generate-billing-invoice.service.js')
-const GenerateBillingInvoiceLicenceService = require('./generate-billing-invoice-licence.service.js')
+const PreGenerateBillingDataService = require('./pre-generate-billing-data.service.js')
 const ProcessBillingTransactionsService = require('./process-billing-transactions.service.js')
 
 /**
@@ -39,7 +38,12 @@ async function go (billingBatch, billingPeriod, chargeVersions) {
 
   const invoiceAccounts = await FetchInvoiceAccountNumbersService.go(chargeVersions)
   const fetchedData = { invoiceAccounts, chargeVersions }
-  const preGeneratedData = _preGenerateData(fetchedData, billingBatchId, billingPeriod)
+  const preGeneratedData = PreGenerateBillingDataService.go(
+    chargeVersions,
+    invoiceAccounts,
+    billingBatchId,
+    billingPeriod
+  )
 
   const billingData = _buildBillingDataWithTransactions(fetchedData, preGeneratedData, billingPeriod, billingBatchId)
   const dataToPersist = await _buildDataToPersist(billingData, billingPeriod, billingBatch)
@@ -47,18 +51,6 @@ async function go (billingBatch, billingPeriod, chargeVersions) {
   await _persistData(dataToPersist)
 
   return dataToPersist.billingInvoiceLicences.length >= 1
-}
-
-/**
- * Pre-generates the empty billing invoices and billing invoice licences we will be populating
- */
-function _preGenerateData (fetchedData, billingBatchId, billingPeriod) {
-  const { chargeVersions, invoiceAccounts } = fetchedData
-
-  const billingInvoices = _preGenerateBillingInvoices(invoiceAccounts, billingBatchId, billingPeriod)
-  const billingInvoiceLicences = _preGenerateBillingInvoiceLicences(chargeVersions, billingInvoices)
-
-  return { billingInvoices, billingInvoiceLicences }
 }
 
 /**
@@ -168,69 +160,8 @@ function _retrievePreGeneratedData (preGeneratedData, chargeVersion) {
   return { billingInvoice, billingInvoiceLicence }
 }
 
-/**
-  * We pre-generate billing invoice licences for every combination of billing invoice and licence in the charge versions
-  * so that we don't need to fetch any data from the db during the main charge version processing loop. This function
-  * generates the required billing invoice licences and returns an object where each key is a concatenated billing
-  * invoice id and licence id, and each value is the billing invoice licence for that combination of billing invoice and
-  * licence, ie:
-  *
-  * {
-  *   'key-1': { billingInvoiceLicenceId: 'billing-invoice-licence-1', ... },
-  *   'key-2': { billingInvoiceLicenceId: 'billing-invoice-licence-2', ... }
-  * }
-  */
-function _preGenerateBillingInvoiceLicences (chargeVersions, billingInvoices) {
-  const keyedBillingInvoiceLicences = chargeVersions.reduce((acc, chargeVersion) => {
-    const { billingInvoiceId } = billingInvoices[chargeVersion.invoiceAccountId]
-    const { licence } = chargeVersion
-
-    const key = _billingInvoiceLicenceKey(billingInvoiceId, licence.licenceId)
-
-    // The charge versions may contain a combination of billing invoice and licence multiple times, so we check to see
-    // if this combination has already had a billing invoice licence generated for it and return early if so
-    if (acc.key) {
-      return acc
-    }
-
-    return {
-      ...acc,
-      [key]: GenerateBillingInvoiceLicenceService.go(billingInvoiceId, licence)
-    }
-  }, {})
-
-  return keyedBillingInvoiceLicences
-}
-
 function _billingInvoiceLicenceKey (billingInvoiceId, licenceId) {
   return `${billingInvoiceId}-${licenceId}`
-}
-
-/**
-  * We pre-generate billing invoices for every invoice account so that we don't need to fetch any data from the db
-  * during the main charge version processing loop. This function generates the required billing invoice licences and
-  * returns an object where each key is the invoice account id, and each value is the billing invoice, ie:
-  *
-  * {
-  *   'uuid-1': { invoiceAccountId: 'uuid-1', ... },
-  *   'uuid-2': { invoiceAccountId: 'uuid-2', ... }
-  * }
-  */
-function _preGenerateBillingInvoices (invoiceAccounts, billingBatchId, billingPeriod) {
-  const keyedBillingInvoices = invoiceAccounts.reduce((acc, invoiceAccount) => {
-    // Note that the array of invoice accounts will already have been deduped so we don't need to check whether a
-    // billing invoice licence already exists in the object before generating one
-    return {
-      ...acc,
-      [invoiceAccount.invoiceAccountId]: GenerateBillingInvoiceService.go(
-        invoiceAccount,
-        billingBatchId,
-        billingPeriod.endDate.getFullYear()
-      )
-    }
-  }, {})
-
-  return keyedBillingInvoices
 }
 
 function _initialBillingData (chargeVersion, billingInvoice, billingInvoiceLicence) {
